ai/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python-headless
numpy
pillow
requests
python-multipart
